requirements.txt
google-genai
python-dotenv
pyperclip